+++ libraries/template-haskell/changelog.md
@@ -25,6 +25,9 @@
   * Add `Semigroup` and `Monoid` instances for `Q` (#18123).
 
   * Add `MonadFix` instance for `Q` (#12073).
+  
+  * Add support for QualifiedDo. The data constructors `DoE` and `MDoE` got a new
+    `Maybe ModName` argument to describe the qualifier of do blocks.
 
 ## 2.16.0.0 *TBA*